# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: codexws HTTP API
  version: 3.0.0
  description: 'Current same-origin API for the one shared codexws backend process.

    This contract documents deployed paths; it is not the proposed /api/v1 design.

    '
servers:
- url: /
...
paths:
...
  /api/codex-history:
    get:
      tags:
      - conversations
      operationId: listConversations
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: include_archived
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Canonical Codex conversations for the exact workspace cwd.
          content:
            application/json:
              schema:
                type: object
                required:
                - conversations
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
  /api/codex-history/{thread_id}:
    parameters:
    - $ref: '#/components/parameters/ThreadIdPath'
    patch:
      tags:
      - conversations
      operationId: updateConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationUpdates'
      responses:
        '200':
          description: Updated local conversation metadata.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                properties:
                  ok:
                    type: boolean
                  conversation:
                    $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - conversations
      operationId: updateConversationPost
      deprecated: true
      description: Same canonical update operation for clients unable to issue PATCH.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationUpdates'
      responses:
        '200':
          description: Updated local conversation metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/Error'
...
  /api/codex-history/{thread_id}/location:
    get:
      summary: Locate a local history entry
      responses:
        '200':
          description: Success
        '400':
          description: Invalid request
        '409':
          description: Unavailable or conflicting state
  /api/codex-history/{thread_id}/read:
    get:
      summary: Read a bounded JSONL page without loading a runtime
      responses:
        '200':
          description: Success
        '400':
          description: Invalid request
        '409':
          description: Unavailable or conflicting state
components:
  parameters:
    WorkspaceId:
      name: workspace_id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
    ThreadIdPath:
      name: thread_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...
    Conversation:
      type: object
      required:
      - threadId
      - title
      - timestamp
      additionalProperties: true
      properties:
        threadId:
          type: string
        forkedFromId:
          type:
          - string
          - 'null'
        title:
          type: string
        rawTitle:
          type: string
        customTitle:
          type:
          - string
          - 'null'
        timestamp:
          type: number
        important:
          type: boolean
        archived:
          type: boolean
    ConversationUpdates:
      type: object
      properties:
        title:
          type:
          - string
          - 'null'
          maxLength: 200
        important:
          type: boolean
        archived:
          type: boolean
      minProperties: 1
...